restructured file structure closed #21
