--- test/.eslintrc.json
@@ -1,6 +1,0 @@
-{
-    "extends": "../.eslintrc.json",
-    "env": {
-	"mocha": true
-    }
-}
